Delayed execution of `if` branches.
The `if` function's branched will have code which can execute
side-effects. For example consider the expression:

  if(condition, dispatch_event("submit"), dispatch_event("error"))

In the current implementation both branches would be executed, which is
not something desired.

This change provides special handling for the `if` function since this
is the only function currently with this property. The arguements for
this function are resolved during function execution (similar to
exprefs). Additionally current data value is passed to the function call
as an extra argument in addition to the `unresolvedArgs` argument.